+++ coinlib/lib/src/tx/sighash/witness_signature_hasher.dart
@@ -44,7 +44,7 @@ final class WitnessSignatureHasher extends SignatureHasher with Writable {
     writer.writeUInt64(details.value);
     writer.writeUInt32(thisInput.sequence.value);
     writer.writeSlice(hashOutputs);
-    writer.writeUInt32(tx.locktime);
+    writer.writeUInt32(tx.locktime.value);
     writer.writeUInt32(hashType.value);
 
   }